WIP - autorefunding of channel points if song isnt found.
Need to add webhook for EventSub - it's the cleanest way of grabbing events and means we don't need to handle redeeming with tmi.js. This is also one of the 2 ways we can get the id for the redemption, the other being grabbing all redemptions.
Additionally need to create the reward programatically - otherwise it is not possible to update the redemption status.